refactor(fare): package 변경 리팩토링
- Fare 객체 관련 line -> path 패키지로 이동
- line 미사용 코드 삭제 처리
