+++ src/test/java/nextstep/subway/line/domain/SubwayUserTest.java
@@ -1,5 +1,6 @@
 package nextstep.subway.line.domain;
 
+import nextstep.subway.path.infrastructure.SubwayUser;
 import nextstep.subway.member.domain.Member;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
