+++ src/test/java/nextstep/subway/path/PathMockitoExtensionTest.java
@@ -4,9 +4,12 @@ import nextstep.subway.auth.domain.Stranger;
 import nextstep.subway.line.application.LineService;
 import nextstep.subway.line.domain.*;
 import nextstep.subway.line.dto.LineResponse;
+import nextstep.subway.path.infrastructure.SeoulMetroFare;
+import nextstep.subway.path.infrastructure.SeoulMetroType;
 import nextstep.subway.path.application.PathService;
 import nextstep.subway.path.domain.Path;
 import nextstep.subway.path.domain.PathEdge;
+import nextstep.subway.path.domain.SubwayFare;
 import nextstep.subway.path.infrastructure.JGraphPathFinder;
 import nextstep.subway.station.application.StationService;
 import nextstep.subway.station.domain.Station;
