+++ src/test/java/nextstep/subway/path/PathSpringExtensionTest.java
@@ -2,9 +2,12 @@ package nextstep.subway.path;
 
 import nextstep.subway.auth.domain.Stranger;
 import nextstep.subway.line.domain.*;
+import nextstep.subway.path.infrastructure.SeoulMetroFare;
+import nextstep.subway.path.infrastructure.SeoulMetroType;
 import nextstep.subway.path.application.PathService;
 import nextstep.subway.path.domain.Path;
 import nextstep.subway.path.domain.PathEdge;
+import nextstep.subway.path.domain.SubwayFare;
 import nextstep.subway.path.infrastructure.JGraphPathFinder;
 import nextstep.subway.station.domain.Station;
 import nextstep.subway.station.domain.StationRepository;
